Add an argument server_ip to initializer of Vodacash
This provides a way to specify the remote server ip or domain name that is whitelisted
by vodacom to process CDF payments.
Also defined C2B_Number and B2C_Number attributes set to None to provide in future a way to specify
which C2B and B2C paybill or Agent Number to use when interacting with the vodacom payment gateway.